moved commands folder to src
