--- commands/utility/restart.js
@@ -1,15 +1,0 @@
-const auth = require('../../auth.json');
-
-module.exports = {
-    name: 'restart',
-    args: false,
-    description: 'Restart the bot.',
-    async execute(message, _args, _db, client) {
-        if(message.author.id == auth.dmID) {
-            message.channel.send('Restarting...').then(m => {
-                client.destroy()}).then(() => {
-                  client.login(auth.token);
-                });
-        }
-    }
-}
