+++ src/commands/dnd/time.js
@@ -1,4 +1,4 @@
-const auth = require("../../auth.json");
+const auth = require("../../../auth.json");
 const {
     MessageEmbed
 } = require("discord.js");
